+++ ios/appmetrica_push_plugin/Sources/appmetrica_push_plugin/include/appmetrica_push_plugin/AMPFAppMetricaPushPlugin.h
@@ -1,5 +1,5 @@
 
 #import <Flutter/Flutter.h>
 
-@interface AMPFAppMetricaPushPlugin : NSObject<FlutterPlugin>
+@interface AMPFAppMetricaPushPlugin : NSObject<FlutterPlugin, FlutterApplicationLifeCycleDelegate>
 @end