+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.3.0
 
+- Fix notification click processing.
+
 ## 2.2.0
 
 - Updated the minimum supported version of [appmetrica_plugin](https://pub.dev/packages/appmetrica_plugin) to `3.2.1`.