+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.4.0
 
+- Update AGP version to `8.2.0`.
+
 ## 2.3.0
 
 - Fix notification click processing.