+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.5.0
+
 ## 2.4.0
 
 - Update AGP version to `8.2.0`.